Adjust usage examples and fix MS concepts page
Clear workspaces at end of usage examples. RelectometryBackgroundSubtraction
crashes without this. I guess one of the leftover workspaces from the Discus
usage examples gets used as a default for the input workspace on
RelectometryBackgroundSubtraction and it doesn't like it
Fix concepts page call to DiscusMultipleScatteringCorrection
Fix compiler warnings

+++ Framework/Algorithms/inc/MantidAlgorithms/DiscusMultipleScatteringCorrection.h
@@ -92,7 +92,7 @@ private:
   void prepareCumulativeProbForQ(double kinc, const API::MatrixWorkspace_sptr &PInvOfQ);
   void getXMinMax(const Mantid::API::MatrixWorkspace &ws, double &xmin, double &xmax) const;
   API::MatrixWorkspace_uptr prepareQSQ(double kinc);
-  double getKf(const API::MatrixWorkspace_sptr &SOfQ, const int iW, const double kinc);
+  double getKf(const API::MatrixWorkspace_sptr &SOfQ, const size_t iW, const double kinc);
   std::tuple<double, int> sampleKW(const API::MatrixWorkspace_sptr &SOfQ, Kernel::PseudoRandomNumberGenerator &rng,
                                    const double kinc);
   void prepareStructureFactor();
